+++ .storybook/main.js
@@ -1,7 +1,6 @@
 const path = require('path');
 
 module.exports = {
-  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: [
     {
       name: '@storybook/addon-docs',
@@ -23,6 +22,10 @@ module.exports = {
     },
     'storybook-addon-designs',
   ],
+  features: {
+    postcss: false,
+  },
+  stories: ['../src/**/*.stories.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
   webpackFinal: async (config) => {
     config.module.rules.push({
       test: /\.less$/,