+++ src/filter-picker/components/FilterList/index.tsx
@@ -36,7 +36,7 @@ function FilterList(props: FilterListProps) {
     } else {
       setFilterList(list);
     }
-  }, [list]);
+  }, [list.length]);
 
   const expressChange = (v: FilterValueType, index: number) => {
     const subFilter = [...filterList];
@@ -83,12 +83,7 @@ function FilterList(props: FilterListProps) {
           />
         ))}
       </div>
-      <Button
-        icon={<PlusCircleFilled />}
-        type="text"
-        disabled={filterList.some((ele: FilterValueType) => !ele.values.length)}
-        onClick={addFilter}
-      >
+      <Button icon={<PlusCircleFilled />} type="text" disabled={filterList.length >= 5} onClick={addFilter}>
         添加过滤条件
       </Button>
       <Footer