+++ src/table-card/__tests__/TableCard.test.tsx
@@ -61,4 +61,17 @@ describe('<TableCard/> test', () => {
     fireEvent.click(container.getElementsByClassName('gio-table-card-select-close')[0]);
     expect(onChange).toBeCalledWith([], []);
   });
+
+  test('divider display logic', () => {
+    const { container, rerender } = render(getTableCard(getTableCardConfig(['1'])));
+    expect(container.getElementsByClassName('gio-table-card-divider')).toHaveLength(2);
+    rerender(getTableCard({ ...getTableCardConfig(['1']), title: undefined, showTabs: false }));
+    expect(container.getElementsByClassName('gio-table-card-divider')).toHaveLength(0);
+  });
+
+  test('otherTabs prop', () => {
+    const { getAllByTestId } = render(getTableCard({ otherTabs: [{ name: 'other', content: 'content' }] }));
+    fireEvent.click(getAllByTestId('tabnav-item')[2]);
+    expect(getAllByTestId('tabnav-item')[2].textContent.includes('other')).toBe(true);
+  });
 });