+++ CHANGELOG.md
@@ -1,3 +1,86 @@
+## [21.5.3](https://github.com/growingio/gio-design-pro/compare/v21.5.1...v21.5.3) (2021-05-17)
+
+
+### Bug Fixes
+
+* **selector:** selector参数dropdownRender判断如果是函数,执行渲染返回元素,如果不是函数,直接渲染元素 ([c345ab6](https://github.com/growingio/gio-design-pro/commit/c345ab66f5197d630fe71a5df7e5005091134cf8))
+
+
+### Features
+
+* **selector:** add prop getContainer ([d88d052](https://github.com/growingio/gio-design-pro/commit/d88d0525a5f6d10f939ffea124bb8f37a536191c))
+* **tablecard:** add otherTabs prop ([9e6b115](https://github.com/growingio/gio-design-pro/commit/9e6b115defacd4a2db73b0657fb3112f737d9414))
+
+
+
+## [21.5.1](https://github.com/growingio/gio-design-pro/compare/v21.5.0...v21.5.1) (2021-05-12)
+
+
+### Bug Fixes
+
+* **filter-picker:** fix filterlist reset ([#242](https://github.com/growingio/gio-design-pro/issues/242)) ([d24b7a3](https://github.com/growingio/gio-design-pro/commit/d24b7a33db29313edb692c4f542660cbef7e71db))
+* **filter-picker:** 参数visible&onChange ([#243](https://github.com/growingio/gio-design-pro/issues/243)) ([81cfb85](https://github.com/growingio/gio-design-pro/commit/81cfb8546cc1428f9ea1b6de959239d51c0221cf))
+
+
+### Features
+
+* **filter-picker:** 增加参数placement ([#244](https://github.com/growingio/gio-design-pro/issues/244)) ([d1d7ab5](https://github.com/growingio/gio-design-pro/commit/d1d7ab5430d86f26b900f8ef5cf89913538b8d98))
+
+
+
+# [21.5.0](https://github.com/growingio/gio-design-pro/compare/v21.4.3...v21.5.0) (2021-05-11)
+
+
+### Features
+
+* **tree-card:** transfer data item to tree node ([f13f812](https://github.com/growingio/gio-design-pro/commit/f13f8124165305a6a866b8b9dc08107cbb47aa2a))
+
+
+
+## [21.4.3](https://github.com/growingio/gio-design-pro/compare/v21.4.2...v21.4.3) (2021-04-29)
+
+
+### Bug Fixes
+
+* **filter-picker:** filter-picker添加operationsOption字段 ([#223](https://github.com/growingio/gio-design-pro/issues/223)) ([4c9456e](https://github.com/growingio/gio-design-pro/commit/4c9456e1bb4211c1d33fe07393d3edbe6ec44dfa))
+* **filter-picker:** 修复日期类型,返回值为数字,以及添加op配置字段 ([#214](https://github.com/growingio/gio-design-pro/issues/214)) ([dcb5a7d](https://github.com/growingio/gio-design-pro/commit/dcb5a7d5fc707f2b2f6a35ed4da320db207aaab3))
+* **operation-menu:** 修复因为gio-design版本升级导致的样式问题 ([#212](https://github.com/growingio/gio-design-pro/issues/212)) ([da7c3e7](https://github.com/growingio/gio-design-pro/commit/da7c3e72fa44086c776dbd399005f733811b9bdc))
+* **operation-menu:** 默认button类型错误 ([86a07a5](https://github.com/growingio/gio-design-pro/commit/86a07a5551ba55ec23efd88721f8700fc1768ae3))
+* **selector:** support size context for selector component ([#228](https://github.com/growingio/gio-design-pro/issues/228)) ([46d28c6](https://github.com/growingio/gio-design-pro/commit/46d28c61493493e18bf5f3f2ed65062a6ea5f148))
+* **table-card:** fix divider style error ([4891af7](https://github.com/growingio/gio-design-pro/commit/4891af7346f540c0e3c629a594eeae7fdd1addc3))
+* **tooltip-button:** 修复tooltip disabled状态颜色的问题 ([#226](https://github.com/growingio/gio-design-pro/issues/226)) ([1aba7c3](https://github.com/growingio/gio-design-pro/commit/1aba7c3f3a6f25b482cf07c5ced3c7e913f2ebe5))
+* **tree-card:** fix style error ([#225](https://github.com/growingio/gio-design-pro/issues/225)) ([36b28fc](https://github.com/growingio/gio-design-pro/commit/36b28fc998baa163a3e03219576be53bdb6b4b65))
+* **tree-card:** 修复tree card组件isLeaf判断的问题 ([#215](https://github.com/growingio/gio-design-pro/issues/215)) ([c93a7eb](https://github.com/growingio/gio-design-pro/commit/c93a7eb33e8826d98c7e48e7688f77cfc88fc430))
+
+
+### Features
+
+* **filter-picker:** 增加受控visible参数 ([#221](https://github.com/growingio/gio-design-pro/issues/221)) ([90bd5ae](https://github.com/growingio/gio-design-pro/commit/90bd5ae178ed580a87f4856d0917cd3a1c9c4f7f))
+
+
+
+## [21.4.2](https://github.com/growingio/gio-design-pro/compare/v21.4.1...v21.4.2) (2021-04-22)
+
+
+### Features
+
+* **tablecard:** add showtabs prop ([f748aac](https://github.com/growingio/gio-design-pro/commit/f748aacef0b8811660a5912db2d6a17d24fb181a))
+* **tablecard:** button support tooltip ([#210](https://github.com/growingio/gio-design-pro/issues/210)) ([d1439ac](https://github.com/growingio/gio-design-pro/commit/d1439acead2f066cc54d8c170c86573ef6b211c9))
+* **tooltip-button:** add tooltip button 组件 ([#209](https://github.com/growingio/gio-design-pro/issues/209)) ([d2d9697](https://github.com/growingio/gio-design-pro/commit/d2d9697b6bf24eea9cd09f57b6282b8958a14762))
+* **treecard:** add treecard component ([#200](https://github.com/growingio/gio-design-pro/issues/200)) ([033d1d8](https://github.com/growingio/gio-design-pro/commit/033d1d8c14447b43a16694a61cb4d7134263490f))
+* **user-picker:** add parameters disabledValues and showSegmentCard ([#203](https://github.com/growingio/gio-design-pro/issues/203)) ([b1e9dfe](https://github.com/growingio/gio-design-pro/commit/b1e9dfe51dde1b4eb88bd5aa406e662d52458021))
+
+
+
+## [21.4.1](https://github.com/growingio/gio-design-pro/compare/v21.4.0...v21.4.1) (2021-04-14)
+
+
+### Features
+
+* **title-info-card:** add title-info-card component ([7d50c73](https://github.com/growingio/gio-design-pro/commit/7d50c73bab72e5650ce2bc4f65e34280cc4d63af))
+
+
+
 # [21.4.0](https://github.com/growingio/gio-design-pro/compare/v21.3.4...v21.4.0) (2021-04-02)